+++ src/gebt_poc/solver.cpp
@@ -19,55 +19,5 @@
 
 namespace openturbine::gebt_poc {
 
-void ElementalConstraintForcesGradientMatrix(
-    View1D::const_type gen_coords, View1D::const_type position_vector,
-    View2D constraints_gradient_matrix
-) {
-    auto translation_0 = Kokkos::subview(gen_coords, Kokkos::make_pair(0, 3));
-    auto rotation_0 = Kokkos::subview(gen_coords, Kokkos::make_pair(3, 7));
-    auto rotation_matrix_0 = gen_alpha_solver::EulerParameterToRotationMatrix(rotation_0);
-    auto position_0 = Kokkos::subview(position_vector, Kokkos::make_pair(0, 3));
-
-    // position_cross_prod_matrix = ~{position_0} + ~{translation_0}
-    auto position_0_cross_prod_matrix = gen_alpha_solver::create_cross_product_matrix(position_0);
-    auto translation_0_cross_prod_matrix =
-        gen_alpha_solver::create_cross_product_matrix(translation_0);
-    auto position_cross_prod_matrix = View2D_3x3("position_cross_prod_matrix");
-    Kokkos::deep_copy(position_cross_prod_matrix, position_0_cross_prod_matrix);
-    KokkosBlas::axpy(1., translation_0_cross_prod_matrix, position_cross_prod_matrix);
-
-    // Assemble the constraint gradient matrix i.e. B matrix for the beam element
-    // [B]_6x(n+1) = [
-    //     [B11]_3x3              0            0   ....  0
-    //     [B21]_3x3          [B22]_3x3        0   ....  0
-    // ]
-    // where
-    // [B11]_3x3 = [1]_3x3
-    // [B21]_3x3 = -[rotation_matrix_0]_3x3
-    // [B22]_3x3 = -[rotation_matrix_0]_3x3 * [position_cross_prod_matrix]_3x3
-    // n = order of the element
-    Kokkos::deep_copy(constraints_gradient_matrix, 0.);
-    auto B11 = Kokkos::subview(
-        constraints_gradient_matrix, Kokkos::make_pair(0, 3), Kokkos::make_pair(0, 3)
-    );
-    auto B21 = Kokkos::subview(
-        constraints_gradient_matrix, Kokkos::make_pair(3, 6), Kokkos::make_pair(0, 3)
-    );
-    auto B22 = Kokkos::subview(
-        constraints_gradient_matrix, Kokkos::make_pair(3, 6), Kokkos::make_pair(3, 6)
-    );
-
-    Kokkos::parallel_for(
-        1,
-        KOKKOS_LAMBDA(std::size_t) {
-            B11(0, 0) = 1.;
-            B11(1, 1) = 1.;
-            B11(2, 2) = 1.;
-        }
-    );
-
-    KokkosBlas::scal(B21, -1., rotation_matrix_0);
-    KokkosBlas::gemm("N", "N", -1., rotation_matrix_0, position_cross_prod_matrix, 0., B22);
-}
 
 }  // namespace openturbine::gebt_poc
