+++ src/gebt_poc/solver.h
@@ -8,10 +8,5 @@
 #include "src/utilities/log.h"
 
 namespace openturbine::gebt_poc {
-/// Calculates the constraint gradient matrix for a beam element
-void ElementalConstraintForcesGradientMatrix(
-    View1D::const_type gen_coords, View1D::const_type position_vector,
-    View2D constraints_gradient_matrix
-);
 
 }  // namespace openturbine::gebt_poc
